2 Novas Rotas

+++ ConnectFit-BackEnd/database/migrations/2023_10_08_145842_create_pessoa_usuarios_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pessoa_usuarios', function (Blueprint $table) {
             $table->unsignedInteger("idPessoaUsuario")->primary();
-            $table->tinyInteger('Altura');
+            $table->smallInteger('Altura');
             $table->boolean('Fumante');
             $table->unsignedTinyInteger('idNivelAtiFis');
             $table->unsignedTinyInteger('idObjetivo');
